requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
